pmempool: performance and blk recovery
- fix performance issue with searching for BTT Info structure
- fix issue with checking consistency of a pool which needs recovery

The poor performance whas due to searching for BTT info header in each
and every 4k page, starting right after the pmemblk descriptor. Now the
search is done in pages when the BTT info (and BTT info backup) can be
stored according to defined layout of the BTT structures.

If the pool needs recovery the lba may be duplicated in a flog entry
which was used by last write operation. This is not an error an must be
properly handled.