benchmarks: do not use op_init interface
.. in pmemobj_atomic_lists benchmark.

The 'op_init' interface was used for getting the destination element
for an operation on a list. It was used because getting the random
element was O(N) (traversing through the list). However this operation
can be implemented as O(1) (getting head, tail or middle element
radomly). There is no need to use the 'op_init' because all operations
are O(1) now and we can perform them in 'operation' function.