obj: fix memcheck instrumentation
The original problem was that if root object was allocated from chunks
(huge allocation) the memcheck reported errors:
	Conditional jump or move depends on uninitialised value(s)

in pmemobj_vg_boot() function, because the root object's size was
accessed before registering it's header to valgrind.

This issue doesn't reproduce if root object is allocated from run
because the entire run space was registered as defined in valgrind -
which is wrong because only allocated objects with its headers should
be registered as defined.

The problem also applies to normal objects, because the loop which
registers the objects to valgrind was touching object's headers which
are not registered yet.

This patch fixes above issues by registering allocated objects during
booting the heap and at the same time when heap's metadata is
registered.